refactor: comprehensive codebase cleanup and optimization
## Phase 0: Fix Next.js 15 Metadata Warnings
- Remove deprecated `appleWebApp` from layout.tsx metadata
- Remove `manifest` property (Next.js auto-detects manifest.json)
- Fixes Vercel deployment failures in PR #8 and #9

## Phase 1: Remove Legacy Code (124KB)
- Delete unused `/app/old` route
- Delete legacy `/components/playground` (26 files)
- Remove 120KB of dead code

## Phase 2: Optimize Icon Assets (Save 2.94MB!)
- Optimize icon-192.png: 811KB → 41KB (95% reduction)
- Optimize icon-512.png: 811KB → 214KB (74% reduction)
- Optimize favicon.ico: 811KB → 3KB (99.6% reduction)
- Delete duplicate rodrigues-icon.png (811KB)
- Total savings: 3.2MB → 258KB = 2.94MB reduction (92%)

## Phase 3: Reorganize /lib Structure
- Clean up lib/utils.ts (remove duplicates, add docs)
- Move getJsonMarkdown to lib/utils/format.ts
- Create lib/auth/ directory for better organization:
  - auth-api.ts → lib/auth/api.ts
  - auth-cookies.ts → lib/auth/cookies.ts
  - validations/auth.ts → lib/auth/validations.ts
- Update all imports across 3 files

## Phase 4: Documentation Organization
- Create /docs folder for better structure
- Move README.md and CONTRIBUTING.md to /docs
- Create new concise root README.md
- Add comprehensive docs/ARCHITECTURE.md

## Phase 5: Developer Experience Improvements
- Install husky + lint-staged for code quality
- Add pre-commit hooks (auto-lint + format)
- Add useful scripts: clean, prepare
- Configure lint-staged for TypeScript and Markdown files

## Phase 6: Code Quality
- Remove all TODOs (were in deleted playground code)
- All validation passing: lint ✓ format ✓ typecheck ✓
- Production build successful ✓

**Impact:**
- 🚀 Fixes Vercel deployments
- 📉 Reduces bundle by 3MB+
- 🧹 Removes 124KB legacy code
- ✅ Eliminates metadata warnings
- 🛡️ Adds git hooks for code quality
- 📚 Better documentation structure
