+++ .husky/pre-commit
@@ -1,0 +1,1 @@
+pnpm lint-staged
