+++ public/fallback-xZAC1b7Iy4OTzt1hqS3-X.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";self.fallback=async e=>"document"===e.destination?caches.match("/offline",{ignoreSearch:!0}):Response.error()})();
