+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-import type { Metadata } from 'next'
+import type { Metadata, Viewport } from 'next'
 import { DM_Mono, Geist, Inter } from 'next/font/google'
 import { NuqsAdapter } from 'nuqs/adapters/next/app'
 import { Toaster } from '@/components/ui/sonner'
@@ -26,27 +26,22 @@ const dmMono = DM_Mono({
   weight: '400'
 })
 
+export const viewport: Viewport = {
+  width: 'device-width',
+  initialScale: 1,
+  maximumScale: 5, // Permitir zoom para acessibilidade
+  userScalable: true,
+  viewportFit: 'cover', // iOS notch/dynamic island support
+  themeColor: [
+    { media: '(prefers-color-scheme: light)', color: '#ffffff' },
+    { media: '(prefers-color-scheme: dark)', color: '#0a0a0a' }
+  ]
+}
+
 export const metadata: Metadata = {
   title: 'Rodrigues AI - Especialista em Crédito Agro e CPR',
   description:
     'Consulte o Rodrigues AI, especialista em crédito agrícola e CPR (Cédula de Produto Rural). Obtenha orientações especializadas sobre financiamento rural, instrumentos de crédito e mercado agropecuário.',
-  viewport: {
-    width: 'device-width',
-    initialScale: 1,
-    maximumScale: 5, // Permitir zoom para acessibilidade
-    userScalable: true,
-    viewportFit: 'cover' // iOS notch/dynamic island support
-  },
-  manifest: '/manifest.json', // PWA manifest
-  appleWebApp: {
-    capable: true,
-    title: 'Rodrigues AI',
-    statusBarStyle: 'black-translucent'
-  },
-  themeColor: [
-    { media: '(prefers-color-scheme: light)', color: '#ffffff' },
-    { media: '(prefers-color-scheme: dark)', color: '#0a0a0a' }
-  ],
   icons: {
     icon: [
       { url: '/icon-192.png', sizes: '192x192', type: 'image/png' },
