--- src/components/playground/ChatArea/ChatInput/index.ts
@@ -1,3 +1,0 @@
-import ChatInput from './ChatInput'
-
-export default ChatInput
