--- src/components/playground/ChatArea/MessageArea.tsx
@@ -1,27 +1,0 @@
-'use client'
-
-import { usePlaygroundStore } from '@/store'
-import Messages from './Messages'
-import ScrollToBottom from '@/components/playground/ChatArea/ScrollToBottom'
-import { StickToBottom } from 'use-stick-to-bottom'
-
-const MessageArea = () => {
-  const { messages } = usePlaygroundStore()
-
-  return (
-    <StickToBottom
-      className="relative mb-4 flex flex-grow flex-col overflow-auto"
-      resize="smooth"
-      initial="smooth"
-    >
-      <StickToBottom.Content className="flex min-h-full flex-col justify-center">
-        <div className="mx-auto w-full max-w-2xl space-y-9 px-2 px-4 pb-4 md:px-4">
-          <Messages messages={messages} />
-        </div>
-      </StickToBottom.Content>
-      <ScrollToBottom />
-    </StickToBottom>
-  )
-}
-
-export default MessageArea
