--- src/components/playground/ChatArea/Messages/Multimedia/Audios/index.ts
@@ -1,3 +1,0 @@
-import Audios from './Audios'
-
-export default Audios
