--- src/components/playground/ChatArea/Messages/Multimedia/Images/index.ts
@@ -1,3 +1,0 @@
-import Images from './Images'
-
-export default Images
