--- src/components/playground/ChatArea/Messages/Multimedia/Videos/index.ts
@@ -1,3 +1,0 @@
-import Videos from './Videos'
-
-export default Videos
