--- src/components/playground/ChatArea/Messages/index.ts
@@ -1,3 +1,0 @@
-import Messages from './Messages'
-
-export default Messages
