--- src/components/playground/ChatArea/index.ts
@@ -1,3 +1,0 @@
-import ChatArea from './ChatArea'
-
-export { ChatArea }
