--- src/components/playground/Sidebar/Sessions/index.ts
@@ -1,3 +1,0 @@
-import Sessions from './Sessions'
-
-export default Sessions
